+++ home_manager/module.nix
@@ -22,6 +22,7 @@ let
   ];
   runtimeVariants = [
     "ghostty"
+    "kitty"
     "yzxterm"
     "wezterm"
   ] ++ lib.optional pkgs.stdenv.hostPlatform.isLinux "ratty";
@@ -29,6 +30,8 @@ let
     runtimeVariant:
     if runtimeVariant == "ghostty" then
       if pkgs.stdenv.hostPlatform.isDarwin then pkgs."ghostty-bin" else pkgs.ghostty
+    else if runtimeVariant == "kitty" then
+      pkgs.kitty
     else if runtimeVariant == "wezterm" then
       pkgs.wezterm
     else if runtimeVariant == "ratty" then
@@ -53,6 +56,12 @@ let
         if runtimeVariant == "wezterm" then
           [
             "ghostty"
+          ]
+        else if runtimeVariant == "kitty" then
+          [
+            "ghostty"
+            "yzxterm"
+            "wezterm"
           ]
         else if runtimeVariant == "ratty" then
           [
@@ -471,6 +480,7 @@ in
         Packaged terminal runtime variant.
 
         - "ghostty": default packaged runtime with Yazelix cursor trails, Ghostty config effects, and Yazi image previews through Zellij
+        - "kitty": packaged Kitty runtime with generated Kitty config and the Yazelix Zellij/Yazi Kitty graphics bridge
         - "wezterm": explicit alternate packaged runtime
         - "yzxterm": experimental Yazelix-owned Rio fork with Ghostty-style cursor shaders
         - "ratty": experimental Linux packaged runtime with Ratty and the Yazelix Zellij/Yazi Kitty graphics bridge